Adding ability to synchronise accounts into firebae

+++ core-coinbase-nonpro/src/main/kotlin/com/dllewellyn/coinbaseapi/models/OauthProvider.kt
@@ -2,8 +2,8 @@ package com.dllewellyn.coinbaseapi.models
 
 data class OauthProvider(
     val access_token: String,
-    val created_at: Int,
-    val expires_in: Int,
+    val created_at: Double,
+    val expires_in: Double,
     val refresh_token: String,
     val scope: String,
     val token_type: String
